Improved stagenode caching

See merge request simulant/simulant!602

+++ simulant/core/stage_node_id.h
@@ -4,7 +4,7 @@
 
 namespace smlt {
 
-typedef uint64_t StageNodeID;
+typedef uint32_t StageNodeID;
 
 bool stage_node_id_matches_type(StageNodeID id, uint16_t node_type);
 StageNodeID new_stage_node_id(uint16_t node_type);
